refactor: package medication together
