+++ HospitalSystem.ConsoleUI/DoctorUI.cs
@@ -3,6 +3,8 @@ using MongoDB.Driver;
 using HospitalSystem.Core.Utils;
 using HospitalSystem.Core;
 using HospitalSystem.Core.Equipment;
+using HospitalSystem.Core.Medications;
+using HospitalSystem.Core.Medications.Requests;
 
 namespace HospitalSystem.ConsoleUI;
