+++ HospitalSystem.Core/Medications/Medication.cs
@@ -1,7 +1,7 @@
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Medications;
 
 public class Medication
 {
