+++ HospitalSystem.Core/Medications/Requests/MedicationRequestRepository.cs
@@ -1,6 +1,6 @@
 using MongoDB.Driver;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Medications.Requests;
 
 public class MedicationRequestRepository : IMedicationRequestRepository
 {
