+++ HospitalSystem.Core/People/Patient/PatientService.cs
@@ -1,5 +1,6 @@
 using MongoDB.Driver;
 using MongoDB.Bson;
+using HospitalSystem.Core.Medications;
 
 namespace HospitalSystem.Core;
