+++ HospitalSystem.Core/People/Patient/Prescription.cs
@@ -1,3 +1,5 @@
+using HospitalSystem.Core.Medications;
+
 namespace HospitalSystem.Core
 {
     public enum MedicationBestTaken
