+++ Tests/TestGenerator.cs
@@ -5,6 +5,8 @@ using HospitalSystem.Core;
 using HospitalSystem.Core.Rooms;
 using HospitalSystem.Core.Equipment;
 using HospitalSystem.Core.Surveys;
+using HospitalSystem.Core.Medications;
+using HospitalSystem.Core.Medications.Requests;
 using HospitalSystem.Core.Utils;
 
 public static class TestGenerator
